fix(hub-discussions): properly export types and interfaces
affects: @esri/hub-discussions
