+++ packages/discussions/src/utils/channels/index.ts
@@ -1,7 +1,7 @@
 import { IUser } from "@esri/arcgis-rest-auth";
 import { GroupMembership } from "@esri/arcgis-rest-portal";
 import { IChannel } from "../../channels";
-import { IPlatformSharing } from "../../types";
+import { IPlatformSharing } from "../../shared/types";
 import { isOrgAdmin, reduceByGroupMembership } from "../platform";
 
 function intersectGroups(
